+++ scm/domain-base/main/src/api/base/errors/mod.rs
@@ -1,0 +1,5 @@
+pub mod request_error;
+pub mod response_error;
+
+pub use request_error::RequestError;
+pub use response_error::ResponseError;