+++ scm/domain-base/main/src/api/base/mod.rs
@@ -1,4 +1,8 @@
+pub mod dto;
+pub mod errors;
 pub mod traits;
 
+pub use dto::{ValidationRequest, ValidationResponse};
+pub use errors::{RequestError, ResponseError};
 pub use traits::Request;
 pub use traits::Response;