+++ scm/domain-base/main/src/api/mod.rs
@@ -1,4 +1,8 @@
 mod base;
 
 pub use base::Request;
+pub use base::RequestError;
 pub use base::Response;
+pub use base::ResponseError;
+pub use base::ValidationRequest;
+pub use base::ValidationResponse;